+++ queries/html/rainbow-tags.scm
@@ -5,3 +5,8 @@
   ((start_tag) @opening
    ; (element (self_closing_tag) @intermediate)?
    (end_tag) @closing)) @container
+
+(style_element
+  ((start_tag) @opening
+   ; (element (self_closing_tag) @intermediate)?
+   (end_tag) @closing)) @container